Safari-611.1.10/LayoutTests/svg/animations/additive-type-by-animation-expected.txt

SVG 1.1 dynamic animation tests


This by animation for all XML property types

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS marker.orientAngle.animVal.value is -45
PASS feConvolveMatrix.divisor.animVal is 37.5
PASS feConvolveMatrix.orderX.animVal is 6
PASS feConvolveMatrix.orderY.animVal is 6
PASS feConvolveMatrix.targetX.animVal is 5
PASS feConvolveMatrix.kernelUnitLengthX.animVal is 20
PASS feConvolveMatrix.kernelUnitLengthY.animVal is 30
PASS feConvolveMatrix.kernelMatrix.animVal.numberOfItems is 9
PASS feConvolveMatrix.kernelMatrix.animVal.getItem(0).value is 0
PASS feConvolveMatrix.kernelMatrix.animVal.getItem(1).value is 1
PASS feConvolveMatrix.kernelMatrix.animVal.getItem(2).value is 0
PASS feConvolveMatrix.kernelMatrix.animVal.getItem(3).value is 0
PASS feConvolveMatrix.kernelMatrix.animVal.getItem(4).value is 1
PASS feConvolveMatrix.kernelMatrix.animVal.getItem(5).value is 0
PASS feConvolveMatrix.kernelMatrix.animVal.getItem(6).value is 0
PASS feConvolveMatrix.kernelMatrix.animVal.getItem(7).value is 1
PASS feConvolveMatrix.kernelMatrix.animVal.getItem(8).value is 0
PASS rect.y.animVal.value is 0
PASS text.dy.animVal.numberOfItems is 4
PASS text.dy.animVal.getItem(0).value is 5
PASS text.dy.animVal.getItem(1).value is -10
PASS text.dy.animVal.getItem(2).value is 10
PASS text.dy.animVal.getItem(3).value is -10
PASS svg.viewBox.animVal.x is 0
PASS svg.viewBox.animVal.y is 0
PASS svg.viewBox.animVal.width is 300
PASS svg.viewBox.animVal.height is 300
PASS polyline.animatedPoints.numberOfItems is 4
PASS polyline.animatedPoints.getItem(0).x is 0
PASS polyline.animatedPoints.getItem(0).y is 0
PASS polyline.animatedPoints.getItem(1).x is 10
PASS polyline.animatedPoints.getItem(1).y is 5
PASS polyline.animatedPoints.getItem(1).x is 10
PASS polyline.animatedPoints.getItem(1).y is 5
PASS polyline.animatedPoints.getItem(2).x is 0
PASS polyline.animatedPoints.getItem(2).y is 10
PASS path.transform.animVal.numberOfItems is 2
PASS path.transform.animVal.getItem(0).type is SVGTransform.SVG_TRANSFORM_ROTATE
PASS path.transform.animVal.getItem(0).angle is 45
PASS path.transform.animVal.getItem(1).type is SVGTransform.SVG_TRANSFORM_ROTATE
PASS path.transform.animVal.getItem(1).angle is 0
PASS colorComponents[1] is 0
PASS colorComponents[2] is 0
PASS colorComponents[3] is 0
PASS marker.orientAngle.baseVal.value is -45
PASS feConvolveMatrix.divisor.baseVal is 37.5
PASS feConvolveMatrix.orderX.baseVal is 6
PASS feConvolveMatrix.orderY.baseVal is 6
PASS feConvolveMatrix.targetX.baseVal is 5
PASS feConvolveMatrix.kernelUnitLengthX.baseVal is 20
PASS feConvolveMatrix.kernelUnitLengthY.baseVal is 30
PASS feConvolveMatrix.kernelMatrix.baseVal.numberOfItems is 9
PASS feConvolveMatrix.kernelMatrix.baseVal.getItem(0).value is 0
PASS feConvolveMatrix.kernelMatrix.baseVal.getItem(1).value is 1
PASS feConvolveMatrix.kernelMatrix.baseVal.getItem(2).value is 0
PASS feConvolveMatrix.kernelMatrix.baseVal.getItem(3).value is 0
PASS feConvolveMatrix.kernelMatrix.baseVal.getItem(4).value is 1
PASS feConvolveMatrix.kernelMatrix.baseVal.getItem(5).value is 0
PASS feConvolveMatrix.kernelMatrix.baseVal.getItem(6).value is 0
PASS feConvolveMatrix.kernelMatrix.baseVal.getItem(7).value is 1
PASS feConvolveMatrix.kernelMatrix.baseVal.getItem(8).value is 0
PASS rect.y.baseVal.value is 0
PASS text.dy.baseVal.numberOfItems is 4
PASS text.dy.baseVal.getItem(0).value is 5
PASS text.dy.baseVal.getItem(1).value is -10
PASS text.dy.baseVal.getItem(2).value is 10
PASS text.dy.baseVal.getItem(3).value is -10
PASS svg.viewBox.baseVal.x is 0
PASS svg.viewBox.baseVal.y is 0
PASS svg.viewBox.baseVal.width is 300
PASS svg.viewBox.baseVal.height is 300
PASS polyline.points.numberOfItems is 4
PASS polyline.points.getItem(0).x is 0
PASS polyline.points.getItem(0).y is 0
PASS polyline.points.getItem(1).x is 10
PASS polyline.points.getItem(1).y is 5
PASS polyline.points.getItem(1).x is 10
PASS polyline.points.getItem(1).y is 5
PASS polyline.points.getItem(2).x is 0
PASS polyline.points.getItem(2).y is 10
PASS path.transform.baseVal.numberOfItems is 1
PASS path.transform.baseVal.getItem(0).type is SVGTransform.SVG_TRANSFORM_ROTATE
PASS path.transform.baseVal.getItem(0).angle is 45
PASS marker.orientAngle.animVal.value is -22.5
PASS feConvolveMatrix.divisor.animVal is 28.75
PASS feConvolveMatrix.orderX.animVal is 5
PASS feConvolveMatrix.orderY.animVal is 5
PASS feConvolveMatrix.targetX.animVal is 3
PASS feConvolveMatrix.kernelUnitLengthX.animVal is 15
PASS feConvolveMatrix.kernelUnitLengthY.animVal is 20
PASS feConvolveMatrix.kernelMatrix.animVal.numberOfItems is 9
PASS feConvolveMatrix.kernelMatrix.animVal.getItem(0).value is 1
PASS feConvolveMatrix.kernelMatrix.animVal.getItem(1).value is 1.5
PASS feConvolveMatrix.kernelMatrix.animVal.getItem(2).value is 1.5
PASS feConvolveMatrix.kernelMatrix.animVal.getItem(3).value is 1
PASS feConvolveMatrix.kernelMatrix.animVal.getItem(4).value is 1.5
PASS feConvolveMatrix.kernelMatrix.animVal.getItem(5).value is 1.5
PASS feConvolveMatrix.kernelMatrix.animVal.getItem(6).value is 1
PASS feConvolveMatrix.kernelMatrix.animVal.getItem(7).value is 1.5
PASS feConvolveMatrix.kernelMatrix.animVal.getItem(8).value is 1.5
PASS rect.y.animVal.value is 50
PASS text.dy.animVal.numberOfItems is 4
PASS text.dy.animVal.getItem(0).value is 0
PASS text.dy.animVal.getItem(1).value is 0
PASS text.dy.animVal.getItem(2).value is 0
PASS text.dy.animVal.getItem(3).value is 0
PASS svg.viewBox.animVal.x is 0
PASS svg.viewBox.animVal.y is 0
PASS svg.viewBox.animVal.width is 250
PASS svg.viewBox.animVal.height is 250
PASS polyline.animatedPoints.numberOfItems is 4
PASS polyline.animatedPoints.getItem(0).x is 0
PASS polyline.animatedPoints.getItem(0).y is 0
PASS polyline.animatedPoints.getItem(1).x is 15
PASS polyline.animatedPoints.getItem(1).y is 7.5
PASS polyline.animatedPoints.getItem(1).x is 15
PASS polyline.animatedPoints.getItem(1).y is 7.5
PASS polyline.animatedPoints.getItem(2).x is 0
PASS polyline.animatedPoints.getItem(2).y is 15
PASS path.transform.animVal.numberOfItems is 2
PASS path.transform.animVal.getItem(0).type is SVGTransform.SVG_TRANSFORM_ROTATE
PASS path.transform.animVal.getItem(0).angle is 45
PASS path.transform.animVal.getItem(1).type is SVGTransform.SVG_TRANSFORM_ROTATE
PASS path.transform.animVal.getItem(1).angle is -22.5
PASS colorComponents[1] is 0
PASS colorComponents[2] is 63
PASS colorComponents[3] is 0
PASS marker.orientAngle.baseVal.value is -45
PASS feConvolveMatrix.divisor.baseVal is 37.5
PASS feConvolveMatrix.orderX.baseVal is 6
PASS feConvolveMatrix.orderY.baseVal is 6
PASS feConvolveMatrix.targetX.baseVal is 5
PASS feConvolveMatrix.kernelUnitLengthX.baseVal is 20
PASS feConvolveMatrix.kernelUnitLengthY.baseVal is 30
PASS feConvolveMatrix.kernelMatrix.baseVal.numberOfItems is 9
PASS feConvolveMatrix.kernelMatrix.baseVal.getItem(0).value is 0
PASS feConvolveMatrix.kernelMatrix.baseVal.getItem(1).value is 1
PASS feConvolveMatrix.kernelMatrix.baseVal.getItem(2).value is 0
PASS feConvolveMatrix.kernelMatrix.baseVal.getItem(3).value is 0
PASS feConvolveMatrix.kernelMatrix.baseVal.getItem(4).value is 1
PASS feConvolveMatrix.kernelMatrix.baseVal.getItem(5).value is 0
PASS feConvolveMatrix.kernelMatrix.baseVal.getItem(6).value is 0
PASS feConvolveMatrix.kernelMatrix.baseVal.getItem(7).value is 1
PASS feConvolveMatrix.kernelMatrix.baseVal.getItem(8).value is 0
PASS rect.y.baseVal.value is 0
PASS text.dy.baseVal.numberOfItems is 4
PASS text.dy.baseVal.getItem(0).value is 5
PASS text.dy.baseVal.getItem(1).value is -10
PASS text.dy.baseVal.getItem(2).value is 10
PASS text.dy.baseVal.getItem(3).value is -10
PASS svg.viewBox.baseVal.x is 0
PASS svg.viewBox.baseVal.y is 0
PASS svg.viewBox.baseVal.width is 300
PASS svg.viewBox.baseVal.height is 300
PASS polyline.points.numberOfItems is 4
PASS polyline.points.getItem(0).x is 0
PASS polyline.points.getItem(0).y is 0
PASS polyline.points.getItem(1).x is 10
PASS polyline.points.getItem(1).y is 5
PASS polyline.points.getItem(1).x is 10
PASS polyline.points.getItem(1).y is 5
PASS polyline.points.getItem(2).x is 0
PASS polyline.points.getItem(2).y is 10
PASS path.transform.baseVal.numberOfItems is 1
PASS path.transform.baseVal.getItem(0).type is SVGTransform.SVG_TRANSFORM_ROTATE
PASS path.transform.baseVal.getItem(0).angle is 45
PASS marker.orientAngle.animVal.value is 0
PASS feConvolveMatrix.divisor.animVal is 20
PASS feConvolveMatrix.orderX.animVal is 3
PASS feConvolveMatrix.orderY.animVal is 3
PASS feConvolveMatrix.targetX.animVal is 1
PASS feConvolveMatrix.kernelUnitLengthX.animVal is 10
PASS feConvolveMatrix.kernelUnitLengthY.animVal is 10
PASS feConvolveMatrix.kernelMatrix.animVal.numberOfItems is 9
PASS feConvolveMatrix.kernelMatrix.animVal.getItem(0).value is 2
PASS feConvolveMatrix.kernelMatrix.animVal.getItem(1).value is 2
PASS feConvolveMatrix.kernelMatrix.animVal.getItem(2).value is 3
PASS feConvolveMatrix.kernelMatrix.animVal.getItem(3).value is 2
PASS feConvolveMatrix.kernelMatrix.animVal.getItem(4).value is 2
PASS feConvolveMatrix.kernelMatrix.animVal.getItem(5).value is 3
PASS feConvolveMatrix.kernelMatrix.animVal.getItem(6).value is 2
PASS feConvolveMatrix.kernelMatrix.animVal.getItem(7).value is 2
PASS feConvolveMatrix.kernelMatrix.animVal.getItem(8).value is 3
PASS rect.y.animVal.value is 100
PASS text.dy.animVal.numberOfItems is 4
PASS text.dy.animVal.getItem(0).value is -5
PASS text.dy.animVal.getItem(1).value is 10
PASS text.dy.animVal.getItem(2).value is -10
PASS text.dy.animVal.getItem(3).value is 10
PASS svg.viewBox.animVal.x is 0
PASS svg.viewBox.animVal.y is 0
PASS svg.viewBox.animVal.width is 200
PASS svg.viewBox.animVal.height is 200
PASS polyline.animatedPoints.numberOfItems is 4
PASS polyline.animatedPoints.getItem(0).x is 0
PASS polyline.animatedPoints.getItem(0).y is 0
PASS polyline.animatedPoints.getItem(1).x is 20
PASS polyline.animatedPoints.getItem(1).y is 10
PASS polyline.animatedPoints.getItem(1).x is 20
PASS polyline.animatedPoints.getItem(1).y is 10
PASS polyline.animatedPoints.getItem(2).x is 0
PASS polyline.animatedPoints.getItem(2).y is 20
PASS path.transform.animVal.numberOfItems is 2
PASS path.transform.animVal.getItem(0).type is SVGTransform.SVG_TRANSFORM_ROTATE
PASS path.transform.animVal.getItem(0).angle is 45
PASS path.transform.animVal.getItem(1).type is SVGTransform.SVG_TRANSFORM_ROTATE
PASS path.transform.animVal.getItem(1).angle is -45
PASS colorComponents[1] is 0
PASS colorComponents[2] is 128
PASS colorComponents[3] is 0
PASS marker.orientAngle.baseVal.value is -45
PASS feConvolveMatrix.divisor.baseVal is 37.5
PASS feConvolveMatrix.orderX.baseVal is 6
PASS feConvolveMatrix.orderY.baseVal is 6
PASS feConvolveMatrix.targetX.baseVal is 5
PASS feConvolveMatrix.kernelUnitLengthX.baseVal is 20
PASS feConvolveMatrix.kernelUnitLengthY.baseVal is 30
PASS feConvolveMatrix.kernelMatrix.baseVal.numberOfItems is 9
PASS feConvolveMatrix.kernelMatrix.baseVal.getItem(0).value is 0
PASS feConvolveMatrix.kernelMatrix.baseVal.getItem(1).value is 1
PASS feConvolveMatrix.kernelMatrix.baseVal.getItem(2).value is 0
PASS feConvolveMatrix.kernelMatrix.baseVal.getItem(3).value is 0
PASS feConvolveMatrix.kernelMatrix.baseVal.getItem(4).value is 1
PASS feConvolveMatrix.kernelMatrix.baseVal.getItem(5).value is 0
PASS feConvolveMatrix.kernelMatrix.baseVal.getItem(6).value is 0
PASS feConvolveMatrix.kernelMatrix.baseVal.getItem(7).value is 1
PASS feConvolveMatrix.kernelMatrix.baseVal.getItem(8).value is 0
PASS rect.y.baseVal.value is 0
PASS text.dy.baseVal.numberOfItems is 4
PASS text.dy.baseVal.getItem(0).value is 5
PASS text.dy.baseVal.getItem(1).value is -10
PASS text.dy.baseVal.getItem(2).value is 10
PASS text.dy.baseVal.getItem(3).value is -10
PASS svg.viewBox.baseVal.x is 0
PASS svg.viewBox.baseVal.y is 0
PASS svg.viewBox.baseVal.width is 300
PASS svg.viewBox.baseVal.height is 300
PASS polyline.points.numberOfItems is 4
PASS polyline.points.getItem(0).x is 0
PASS polyline.points.getItem(0).y is 0
PASS polyline.points.getItem(1).x is 10
PASS polyline.points.getItem(1).y is 5
PASS polyline.points.getItem(1).x is 10
PASS polyline.points.getItem(1).y is 5
PASS polyline.points.getItem(2).x is 0
PASS polyline.points.getItem(2).y is 10
PASS path.transform.baseVal.numberOfItems is 1
PASS path.transform.baseVal.getItem(0).type is SVGTransform.SVG_TRANSFORM_ROTATE
PASS path.transform.baseVal.getItem(0).angle is 45
PASS successfullyParsed is true

TEST COMPLETE
